notebooks/pyproject.toml
[project]
name = "notebook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.91.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
